# Gaming.Quiz.API/wwwroot/config/swagger/Gaming.Quiz.API.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gaming.Quiz.API",
    "description": "The API documentation for Gaming Quiz APIs",
    "version": "V1"
  },
  "paths": {
...
    "/services/Leaderboard/{userguid}/userrank": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch User Rank on leaderboard",
        "parameters": [
          {
            "name": "OptType",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "MonthId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "GamedayId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "WeekId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          },
          {
            "name": "userguid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/{userguid}/userteamrank": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch User Player Rank on leaderboard",
        "parameters": [
          {
            "name": "OptType",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "PlayerId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          },
          {
            "name": "userguid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/{userguid}/bestscoreuserrank": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch Best score User Rank on leaderboard",
        "parameters": [
          {
            "name": "OptType",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "MonthId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "GamedayId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "WeekId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          },
          {
            "name": "userguid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/overallleaderboard": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch overall Leaderboard",
        "parameters": [
          {
            "name": "OptType",
            "in": "query",
            "description": "OptType - 1",
            "schema": {
              "type": "integer",
              "description": "OptType - 1",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/teamleaderboard": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch Player overall Leaderboard",
        "parameters": [
          {
            "name": "PlayerId",
            "in": "query",
            "description": "MonthId",
            "schema": {
              "type": "integer",
              "description": "MonthId",
              "format": "int64"
            }
          },
          {
            "name": "OptType",
            "in": "query",
            "description": "OptType - 1",
            "schema": {
              "type": "integer",
              "description": "OptType - 1",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/monthleaderboard": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch monthly Leaderboard",
        "parameters": [
          {
            "name": "MonthId",
            "in": "query",
            "description": "MonthId",
            "schema": {
              "type": "integer",
              "description": "MonthId",
              "format": "int64"
            }
          },
          {
            "name": "OptType",
            "in": "query",
            "description": "OptType - 1",
            "schema": {
              "type": "integer",
              "description": "OptType - 1",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/bestscoreleaderboard": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch gloabl Leaderboard",
        "parameters": [
          {
            "name": "MonthId",
            "in": "query",
            "description": "MonthId",
            "schema": {
              "type": "integer",
              "description": "MonthId",
              "format": "int64"
            }
          },
          {
            "name": "OptType",
            "in": "query",
            "description": "OptType - 1",
            "schema": {
              "type": "integer",
              "description": "OptType - 1",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
    "/services/Leaderboard/getmonth": {
      "get": {
        "tags": [
          "Leaderboard"
        ],
        "summary": "API to fetch User Rank on leaderboard",
        "parameters": [
          {
            "name": "backdoor",
            "in": "query",
            "description": "Debug param",
            "schema": {
              "type": "string",
              "description": "Debug param",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FeedTime": {
        "type": "object",
        "properties": {
          "utcTime": {
            "type": "string",
            "nullable": true
          },
          "istTime": {
            "type": "string",
            "nullable": true
          },
          "cestTime": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HTTPMeta": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "retVal": {
            "type": "integer",
            "format": "int64"
          },
          "success": {
            "type": "boolean"
          },
          "timestamp": {
            "$ref": "#/components/schemas/FeedTime"
          }
        },
        "additionalProperties": false
      },
      "HTTPResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": false,
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/HTTPMeta"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}